move code to analysis module; create seperate sentiment function
